pyproject.toml
[project]
name = "vms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apitofsim",
    "duckdb>=1.4.1",
    "holoviews>=1.21.0",
    "lxml>=6.0.2",
    "matplotlib>=3.10.6",
    "minify-html-onepass>=0.16.4",
    "pint>=0.25",
    "pip>=25.2",
    "quart>=0.20.0",
    "quart-wtforms>=1.0.3",
    "ray[default]>=2.49.1",
    "virtualenv>=20.34.0",
    "wtforms>=3.2.1",
]

[project.optional-dependencies]
deploy = [
    "hypercorn>=0.17.0",
]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
apitofsim = { git = "https://github.com/VilmaHelsinki/apitofsim.git", branch = "python" }
